requirements.txt
sqlmodel
fastapi
uvicorn
pytest
python-dotenv
pydantic